+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "anno-modding-tools" extension will be documented in 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+- Added file `FileName` checks to `.cfg` files
+
 ## [1.9.1]
 
 - Fixed crash with large XML files